+++ tests/functional/test_lint_imports.py
@@ -30,50 +30,10 @@ sys.path.extend(
         (testpackage_directory, ".externalbrokencontract.ini", cli.EXIT_STATUS_ERROR),
         (multipleroots_directory, ".multiplerootskeptcontract.ini", cli.EXIT_STATUS_SUCCESS),
         (multipleroots_directory, ".multiplerootsbrokencontract.ini", cli.EXIT_STATUS_ERROR),
-        # TOML versions.
-        pytest.param(
-            testpackage_directory,
-            ".setup.toml",
-            cli.EXIT_STATUS_ERROR,
-            marks=pytest.mark.toml_not_installed,
-        ),
-        pytest.param(
-            testpackage_directory,
-            ".setup.toml",
-            cli.EXIT_STATUS_SUCCESS,
-            marks=pytest.mark.toml_installed,
-        ),
-        pytest.param(
-            testpackage_directory,
-            ".customkeptcontract.toml",
-            cli.EXIT_STATUS_ERROR,
-            marks=pytest.mark.toml_not_installed,
-        ),
-        pytest.param(
-            testpackage_directory,
-            ".customkeptcontract.toml",
-            cli.EXIT_STATUS_SUCCESS,
-            marks=pytest.mark.toml_installed,
-        ),
-        pytest.param(
-            testpackage_directory,
-            ".customkeptcontract.toml",
-            cli.EXIT_STATUS_ERROR,
-            marks=pytest.mark.toml_not_installed,
-        ),
-        pytest.param(
-            testpackage_directory,
-            ".customkeptcontract.toml",
-            cli.EXIT_STATUS_SUCCESS,
-            marks=pytest.mark.toml_installed,
-        ),
+        (testpackage_directory, ".setup.toml", cli.EXIT_STATUS_SUCCESS),
+        (testpackage_directory, ".customkeptcontract.toml", cli.EXIT_STATUS_SUCCESS),
         (testpackage_directory, ".externalkeptcontract.ini", cli.EXIT_STATUS_SUCCESS),
-        pytest.param(
-            testpackage_directory,
-            ".externalkeptcontract.toml",
-            cli.EXIT_STATUS_SUCCESS,
-            marks=pytest.mark.toml_installed,
-        ),
+        (testpackage_directory, ".externalkeptcontract.toml", cli.EXIT_STATUS_SUCCESS),
         # Unmatched ignore imports alerting.
         # The return value depends on what this is set to.
         (